requirements.txt
# Base de Datos
psycopg2-binary

# API Backend
fastapi
uvicorn
pydantic
python-dotenv

# Data Engineering & ETL
pandas
requests
bcchapi

# Dashboard Frontend
streamlit
plotly       <-- ¡Esta es la que faltaba!